ai_service/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
tensorflow==2.16.1
scikit-learn==1.4.2
pandas==2.2.2
numpy==1.26.4
joblib==1.4.2
schedule==1.2.1
influxdb-client==1.43.0
supabase>=2.3.0
python-dotenv==1.0.1
pydantic==2.7.1
httpx==0.27.0
